Reorganise hierarchy & architecture
