+++ edtLHD/src/main/java/fr/univtln/lhd/model/entitys/slots/Classroom.java
@@ -1,9 +1,12 @@
-package fr.univtln.lhd.entitys.slots;
+package fr.univtln.lhd.model.entitys.slots;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+/**
+ * Class defining a Classroom
+ */
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @Getter
 public class Classroom {
@@ -11,6 +14,12 @@ public class Classroom {
     private final String name;
     private final String buildingName; //? might not be a string in database ?
 
+    /**
+     * Factory for a classroom
+     * @param name name of the classroom
+     * @param buildingName name of the building for that classroom
+     * @return an instance of Classroom class
+     */
     public static Classroom getInstance(String name, String buildingName) {
         return new Classroom(name, buildingName);
     }
