+++ edtLHD/src/main/java/fr/univtln/lhd/model/entitys/slots/Subject.java
@@ -1,9 +1,12 @@
-package fr.univtln.lhd.entitys.slots;
+package fr.univtln.lhd.model.entitys.slots;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+/**
+ * Class defining a Subject
+ */
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @Getter
 public class Subject {
@@ -11,6 +14,12 @@ public class Subject {
     private final String name;
     private final int hourCountMax; //max number of hours dedicated to this subject
 
+    /**
+     * Factory for a Subject
+     * @param name name of the subject
+     * @param hourCountMax max number of hours dedicated to this subject
+     * @return an instance of Subject
+     */
     public static Subject getInstance(String name, int hourCountMax){
         return new Subject(name, hourCountMax);
     }
