+++ edtLHD/src/main/java/fr/univtln/lhd/model/entitys/user/Student.java
@@ -1,4 +1,4 @@
-package fr.univtln.lhd.entitys;
+package fr.univtln.lhd.model.entitys.user;
 
 import lombok.Getter;
 
@@ -27,4 +27,23 @@ public class Student extends User {
         return new Student(name, fname, email);
     }
 
+    /**
+     * Take an object and return if the student is equal to it
+     * the only important parameter for the test of equality are those of User
+     * @param o an object
+     * @return a boolean
+     */
+    @Override
+    public boolean equals(Object o) {
+        return super.equals(o);
+    }
+
+    /**
+     * HashCode of student, like the equality the only important parameter are those of User
+     * @return hashcode
+     */
+    @Override
+    public int hashCode() {
+        return super.hashCode();
+    }
 }
