--- edtLHD/src/test/java/fr/univtln/lhd/AppTest.java
@@ -1,20 +1,0 @@
-package fr.univtln.lhd;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-import org.junit.jupiter.api.Test;
-
-/**
- * Unit test for simple App.
- */
-public class AppTest 
-{
-    /**
-     * Rigorous Test :-)
-     */
-    @Test
-     void shouldAnswerWithTrue()
-    {
-        assertTrue( true );
-    }
-}
